TTDA-35 Conected Register form to endpoint
