+++ app-ui/src/services/api.js
@@ -1,0 +1,23 @@
+import * as TokenService from './token-store';
+
+export const request = async (
+  path,
+  { method = 'GET', body, auth = false } = {}
+) => {
+  const headers = { 'Content-Type': 'application/json' };
+  if (auth) headers['Authorization'] = `Bearer ${TokenService.getToken()}`;
+
+  try {
+    const res = await fetch(`${path}`, {
+      method,
+      headers,
+      body: body ? JSON.stringify(body) : undefined,
+    });
+    const { token } = res.json();
+
+    TokenService.setToken(token);
+  } catch (err) {
+    console.log(err.message);
+    console.log(err);
+  }
+};
