+++ app-ui/vite.config.js
@@ -12,9 +12,15 @@ export default defineConfig({
     watch: {
       usePolling: true,
     },
-  },
-  test: {
-    environment: 'jsdom',
-    globals: true,
+    proxy: {
+      '/api': {
+        target: 'http://ttd-backend:3000', // 👈 service name, not localhost
+        changeOrigin: true,
+      },
+    },
+    test: {
+      environment: 'jsdom',
+      globals: true,
+    },
   },
 });
